+++ stablecoin-v2/src/actors/hedging_agents.rs
