+++ stablecoin-v2/src/actors/liquidity_providers.rs
