+++ stablecoin-v2/src/actors/mod.rs
@@ -1,0 +1,4 @@
+pub mod hedging_agents;
+pub mod keepers;
+pub mod liquidity_providers;
+pub mod stable_seekers;
