+++ stablecoin-v2/src/economics/fees.rs
