+++ stablecoin-v2/src/economics/math.rs
