+++ stablecoin-v2/src/economics/mod.rs
@@ -1,0 +1,3 @@
+pub mod fees;
+pub mod math;
+pub mod pools;
