+++ stablecoin-v2/src/economics/pools.rs
