+++ stablecoin-v2/src/lib.rs
@@ -3,21 +3,13 @@
 elrond_wasm::imports!();
 elrond_wasm::derive_imports!();
 
-// tokens
-mod hedging_token;
-mod liquidity_token;
-mod stablecoin_token;
-mod token_common;
+mod actors;
+mod economics;
+mod tokens;
 
-// actors
-mod hedging_agents;
-mod keepers;
-mod liquidity_providers;
-
-// misc
-mod fees;
-mod math;
-mod pools;
+use actors::*;
+use economics::*;
+use tokens::*;
 
 // TODO: Add events
 
@@ -33,6 +25,7 @@ pub trait StablecoinV2:
     + pools::PoolsModule
     + price_aggregator_proxy::PriceAggregatorModule
     + stablecoin_token::StablecoinTokenModule
+    + stable_seekers::StableSeekers
     + token_common::TokenCommonModule
 {
     #[init]
@@ -59,8 +52,6 @@ pub trait StablecoinV2:
 
         Ok(())
     }
-
-    // endpoints - owner-only
 
     #[only_owner]
     #[endpoint(addCollateralToWhitelist)]
@@ -122,91 +113,5 @@ pub trait StablecoinV2:
             .clear();
         self.min_max_slippage_percentage(&collateral_id).clear();
         self.collateral_whitelisted(&collateral_id).clear();
-    }
-
-    // endpoints
-
-    #[payable("*")]
-    #[endpoint(sellCollateral)]
-    fn sell_collateral(
-        &self,
-        #[payment_token] payment_token: TokenIdentifier,
-        #[payment_amount] payment_amount: BigUint,
-        min_amount_out: BigUint,
-    ) -> SCResult<()> {
-        self.require_collateral_in_whitelist(&payment_token)?;
-
-        let collateral_value_in_dollars = self.get_collateral_value_in_dollars(&payment_token)?;
-        let transaction_fees_percentage = self.get_mint_transaction_fees_percentage(&payment_token);
-        let fees_amount_in_collateral =
-            self.calculate_percentage_of(&transaction_fees_percentage, &payment_amount);
-        let collateral_amount = &payment_amount - &fees_amount_in_collateral;
-
-        let stablecoin_amount = &collateral_value_in_dollars * &collateral_amount;
-        require!(stablecoin_amount >= min_amount_out, "Below min amount");
-
-        self.update_pool(&payment_token, |pool| {
-            pool.collateral_amount += &collateral_amount;
-            pool.stablecoin_amount += &stablecoin_amount;
-        });
-        self.accumulated_tx_fees(&payment_token)
-            .update(|accumulated_fees| *accumulated_fees += fees_amount_in_collateral);
-
-        let caller = self.blockchain().get_caller();
-        self.mint_and_send_stablecoin(&caller, &stablecoin_amount);
-
-        Ok(())
-    }
-
-    #[payable("*")]
-    #[endpoint(buyCollateral)]
-    fn buy_collateral(
-        &self,
-        #[payment_token] payment_token: TokenIdentifier,
-        #[payment_amount] payment_amount: BigUint,
-        collateral_id: TokenIdentifier,
-        min_amount_out: BigUint,
-    ) -> SCResult<()> {
-        let stablecoin_token_id = self.stablecoin_token_id().get();
-        require!(
-            payment_token == stablecoin_token_id,
-            "May only pay with stablecoins"
-        );
-        self.require_collateral_in_whitelist(&collateral_id)?;
-
-        let collateral_value_in_dollars = self.get_collateral_value_in_dollars(&collateral_id)?;
-        let total_value_in_collateral = &payment_amount / &collateral_value_in_dollars;
-        let transaction_fees_percentage = self.get_burn_transaction_fees_percentage(&collateral_id);
-        let fees_amount_in_collateral =
-            self.calculate_percentage_of(&transaction_fees_percentage, &total_value_in_collateral);
-
-        let collateral_amount = &total_value_in_collateral - &fees_amount_in_collateral;
-        require!(collateral_amount >= min_amount_out, "Below min amount");
-
-        self.update_pool(&collateral_id, |pool| {
-            require!(
-                pool.collateral_amount >= collateral_amount,
-                "Insufficient funds for swap"
-            );
-            require!(
-                pool.stablecoin_amount >= payment_amount,
-                "Too many stablecoins paid"
-            );
-
-            pool.collateral_amount -= &collateral_amount;
-            pool.stablecoin_amount -= &payment_amount;
-
-            Ok(())
-        })?;
-        self.accumulated_tx_fees(&collateral_id)
-            .update(|accumulated_fees| *accumulated_fees += fees_amount_in_collateral);
-
-        self.burn_stablecoin(&payment_amount);
-
-        let caller = self.blockchain().get_caller();
-        self.send()
-            .direct(&caller, &collateral_id, 0, &collateral_amount, &[]);
-
-        Ok(())
     }
 }
