+++ stablecoin-v2/src/tokens/hedging_token.rs
