+++ stablecoin-v2/src/tokens/liquidity_token.rs
