+++ stablecoin-v2/src/tokens/mod.rs
@@ -1,0 +1,4 @@
+pub mod hedging_token;
+pub mod liquidity_token;
+pub mod stablecoin_token;
+pub mod token_common;
