+++ stablecoin-v2/src/tokens/stablecoin_token.rs
