+++ stablecoin-v2/src/tokens/token_common.rs
